+++ src/main/java/com/masai/controller/CustomerControl.java
@@ -2,6 +2,8 @@ package com.masai.controller;
 
 import com.masai.entity.Customer;
 import com.masai.entity.UserType;
+import com.masai.exception.InvalidId;
+import com.masai.exception.NullValueException;
 import com.masai.service.CustomerService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -21,24 +23,48 @@ public class CustomerControl {
     }
     @GetMapping("/{id}")
     public Customer viewCustomerHandler(@PathVariable Integer id) {
-        return customerService.viewCustomer(id);
+        if(id==null){
+            throw new NullValueException("Invalid URI");
+        }
+        else {
+            return customerService.viewCustomer(id);
+        }
     }
 
     @PutMapping("/{id}/{mobile}/{email}")
     public Customer updateCustomerHandler(@PathVariable Integer id,
                                           @PathVariable String mobile, @PathVariable String email) {
-        Customer customer = customerService.updateCustomer(id, mobile, email);
-        return customer;
+        if(id==null){
+            throw new InvalidId("Invalid URI");
+        }else {
+            Customer customer = customerService.updateCustomer(id, mobile, email);
+            if(customer==null){
+                throw new NullValueException("Customer not found");
+            }
+            return customer;
+        }
+
     }
 
     @DeleteMapping("/{id}")
     public String removeCustomerHandler(@PathVariable Integer id) {
-        Customer customer = customerService.removeCustomer(id);
-        return "Deleted " + customer;
+        if(id==null){
+            throw new InvalidId("Invalid URI");
+        }else {
+            Customer customer = customerService.removeCustomer(id);
+            if(customer==null){
+                throw new NullValueException("Customer not found");
+            }
+            return "Deleted " + customer;
+        }
     }
 
     @GetMapping("/all")
     public List<Customer> viewAllCustomersHandler() {
-        return customerService.viewAllCustomers();
+        List<Customer>customers=customerService.viewAllCustomers();
+        if(customers.size()==0){
+            throw new NullValueException("Customer not found");
+        }
+        return customers;
     }
 }
